refactor: move card component out, set up link behaviour
